third_party/blink/web_tests/svg/dynamic-updates/SVGFEDropShadowElement-dom-shadow-opacity-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'flood-opacity' attribute of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.getAttribute('flood-opacity') is "0"
PASS dropShadowElement.getAttribute('flood-opacity') is "0.5"
PASS successfullyParsed is true

TEST COMPLETE
